+++ serum2vital/__init__.py
@@ -1,3 +1,3 @@
 """Convert Xfer Serum 1 and Serum 2 presets into Vital presets."""
 
-__version__ = "0.6.0"
+__version__ = "0.6.1"